Moved CommandStatusWriter to PortableUtil for cross-tool sharing
* Relocated CommandStatusWriter from CommandLine.cs to pwiz.Common.SystemUtil (PortableUtil), pure-BCL
* Parameterized the localized Error hint via static AddErrorMessageHint; Skyline installs it in Program.Main
* Added CommonUtil and TestUtil project references to PortableUtil

See ai/todos/active/TODO-20260623_ospreysharp_console_output.md
