+++ pwiz_tools/Skyline/CommandLine.cs
@@ -4840,145 +4840,6 @@ namespace pwiz.Skyline
         #endregion
     }
 
-    public class CommandStatusWriter : TextWriter
-    {
-        private TextWriter _writer;
-
-        public CommandStatusWriter(TextWriter writer)
-            : base(writer.FormatProvider)
-        {
-            _writer = Synchronized(writer); // Make this thread safe for more predictable console output
-        }
-
-        public bool IsTimeStamped { get; set; }
-
-        public bool IsMemStamped { get; set; }
-
-        public bool IsErrorReported { get; private set; }
-
-        public bool IsVerboseExceptions { get; set; }
-
-        public override Encoding Encoding
-        {
-            get { return _writer.Encoding; }
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (_writer != null)
-            {
-                _writer.Dispose();
-                _writer = null;
-            }
-        }
-
-        public override void Flush()
-        {
-            _writer.Flush();
-        }
-
-        public override void Write(char value)
-        {
-            _writer.Write(value);
-        }
-
-        public override void Write(string value)
-        {
-            _writer.Write(value);
-        }
-
-        public override void WriteLine()
-        {
-            WriteLine(string.Empty);
-        }
-
-        public void WriteException(string formatMessage, string string0, Exception x1, bool lineSeparate = false)
-        {
-            if (!lineSeparate)
-                WriteLine(formatMessage, string0, ExceptionString(x1));
-            else
-            {
-                WriteLine(formatMessage, string0);
-                WriteException(x1);
-            }
-        }
-        public void WriteException(string formatMessage, Exception x, bool lineSeparate = false)
-        {
-            if (string.IsNullOrEmpty(formatMessage))
-                WriteException(x);
-            else if (!lineSeparate)
-                WriteLine(formatMessage, ExceptionString(x));
-            else
-            {
-                WriteLine(formatMessage);
-                WriteException(x);
-            }
-        }
-        public void WriteException(Exception x)
-        {
-            WriteLine(ExceptionString(x));
-        }
-
-        /// <summary>
-        /// Get a string reporting the exception, with information depending on the verbose exception setting.
-        /// </summary>
-        /// <param name="x">Exception to be reported</param>
-        /// <returns>A message reporting the exception</returns>
-        private string ExceptionString(Exception x)
-        {
-            return IsVerboseExceptions ? x.ToString() : x.Message;
-        }
-
-        public override void WriteLine(string value)
-        {
-            var message = new StringBuilder();
-            if (IsTimeStamped)
-                // ReSharper disable LocalizableElement
-                message.Append(DateTime.Now.ToString("[yyyy/MM/dd HH:mm:ss]\t"));
-                // ReSharper restore LocalizableElement
-            if (IsMemStamped)
-            {
-                lock (_writer)
-                {
-                    // This can take long enough that we need to introduce a lock to keep
-                    // output ordered as much as possible
-                    message.Append(MemStamp(GC.GetTotalMemory(false)));
-                    message.Append(MemStamp(Process.GetCurrentProcess().PrivateMemorySize64));
-                }
-            }
-            message.Append(value);
-            _writer.WriteLine(message);
-            Flush();
-
-            if (IsErrorMessage(value))
-            {
-                IsErrorReported = true;
-            }
-        }
-
-        public const string ERROR_MESSAGE_HINT = @"Error:";
-
-        private bool IsErrorMessage(string message)
-        {
-            if (message != null && !IsErrorReported)
-            {
-                return message.StartsWith(ERROR_MESSAGE_HINT, StringComparison.InvariantCulture) ||  // In Skyline-daily any message might not be localized
-                       message.StartsWith(Resources.CommandStatusWriter_WriteLine_Error_,
-                           StringComparison.CurrentCulture);
-            }
-
-            return false;
-        }
-
-        private string MemStamp(long memUsed)
-        {
-            const double mb = 1024 * 1024;
-            // ReSharper disable LocalizableElement
-            return string.Format("{0}\t", Math.Round(memUsed/mb));
-            // ReSharper restore LocalizableElement
-        }
-    }
-
     public class ExportCommandProperties : ExportProperties
     {
         private readonly TextWriter _out;
