+++ pwiz_tools/Skyline/TestUtil/TestUtil.csproj
@@ -241,6 +241,10 @@
       <Project>{13BF2FFB-50A1-4AB1-83A4-5733E36905CE}</Project>
       <Name>CommonUtil</Name>
     </ProjectReference>
+    <ProjectReference Include="..\..\Shared\PortableUtil\PortableUtil.csproj">
+      <Project>{97ECF0B4-0AAA-4593-ADE0-9E8740973AC2}</Project>
+      <Name>PortableUtil</Name>
+    </ProjectReference>
     <ProjectReference Include="..\..\Shared\MSGraph\MSGraph.csproj">
       <Project>{26cfd1ff-f4f7-4f66-b5b4-e686bdb9b34e}</Project>
       <Name>MSGraph</Name>
